# Use counter + time.AfterFunc instead of goroutines