`time.AfterFunc` handles all timers in a single goroutine, so we don't need to allocate a new goroutine for every request.

Benchmark comparison:
```
// old
BenchmarkLimitedHandlerSingleIP-4   	3000000      399 ns/op    108 B/op      2 allocs/op
BenchmarkLimitedHandlerManyIPs-4    	1000000     2438 ns/op    831 B/op      3 allocs/op

// new
BenchmarkLimitedHandlerSingleIP-4   	5000000      395 ns/op    121 B/op      2 allocs/op
BenchmarkLimitedHandlerManyIPs-4    	2000000      520 ns/op    203 B/op      2 allocs/op
```
The effect is more pronounced when there are many concurrent requests and the timeout is long.

More important is the fact that the new implementation will clean up map entries that hit 0, whereas the channel-based implementation grows unbounded.

I also changed the benchmarks to mutate a single request object instead of allocating a new one in the loop body, which was severely distorting the results.